Make an integration tests directory for the integration tests
