+++ tests/src/Integration/ReplaceOpTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grasmash\ComposerScaffold\Tests\Unit;
+namespace Grasmash\ComposerScaffold\Tests\Integration;
 
 use Composer\Composer;
 use Composer\IO\IOInterface;
@@ -8,12 +8,12 @@ use Composer\Package\PackageInterface;
 use Grasmash\ComposerScaffold\Handler;
 use PHPUnit\Framework\TestCase;
 use Grasmash\ComposerScaffold\Tests\Fixtures;
-use Grasmash\ComposerScaffold\Operations\SkipOp;
+use Grasmash\ComposerScaffold\Operations\ReplaceOp;
 
 /**
- * @coversDefaultClass \Grasmash\ComposerScaffold\Operations\SkipOp
+ * @coversDefaultClass \Grasmash\ComposerScaffold\Operations\ReplaceOp
  */
-class SkipOpTest extends TestCase {
+class ReplaceOpTest extends TestCase {
 
   /**
    * @covers ::process
@@ -26,7 +26,9 @@ class SkipOpTest extends TestCase {
 
     $options = [];
 
-    $sut = new SkipOp();
+    $sut = new ReplaceOp();
+    $sut->setSource($source);
+    $sut->setOverwrite(TRUE);
 
     // Assert that there is no target file before we run our test.
     $this->assertFileNotExists($destination->fullPath());
@@ -34,12 +36,16 @@ class SkipOpTest extends TestCase {
     // Test the system under test.
     $sut->process($destination, $fixtures->io(), $options);
 
-    // Assert that the target file was not created.
-    $this->assertFileNotExists($destination->fullPath());
+    // Assert that the target file was created.
+    $this->assertFileExists($destination->fullPath());
+
+    // Assert the target contained the contents from the correct scaffold file.
+    $contents = trim(file_get_contents($destination->fullPath()));
+    $this->assertEquals('# Test version of robots.txt from drupal/core.', $contents);
 
     // Confirm that expected output was written to our io fixture.
     $output = $fixtures->getOutput();
-    $this->assertContains('Skip [web-root]/robots.txt: disabled', $output);
+    $this->assertContains('Copy [web-root]/robots.txt from assets/robots.txt', $output);
   }
 
 }
