+++ tests/src/Integration/SkipOpTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grasmash\ComposerScaffold\Tests\Unit;
+namespace Grasmash\ComposerScaffold\Tests\Integration;
 
 use Composer\Composer;
 use Composer\IO\IOInterface;
@@ -8,12 +8,12 @@ use Composer\Package\PackageInterface;
 use Grasmash\ComposerScaffold\Handler;
 use PHPUnit\Framework\TestCase;
 use Grasmash\ComposerScaffold\Tests\Fixtures;
-use Grasmash\ComposerScaffold\Operations\ReplaceOp;
+use Grasmash\ComposerScaffold\Operations\SkipOp;
 
 /**
- * @coversDefaultClass \Grasmash\ComposerScaffold\Operations\ReplaceOp
+ * @coversDefaultClass \Grasmash\ComposerScaffold\Operations\SkipOp
  */
-class ReplaceOpTest extends TestCase {
+class SkipOpTest extends TestCase {
 
   /**
    * @covers ::process
@@ -26,9 +26,7 @@ class ReplaceOpTest extends TestCase {
 
     $options = [];
 
-    $sut = new ReplaceOp();
-    $sut->setSource($source);
-    $sut->setOverwrite(TRUE);
+    $sut = new SkipOp();
 
     // Assert that there is no target file before we run our test.
     $this->assertFileNotExists($destination->fullPath());
@@ -36,16 +34,12 @@ class ReplaceOpTest extends TestCase {
     // Test the system under test.
     $sut->process($destination, $fixtures->io(), $options);
 
-    // Assert that the target file was created.
-    $this->assertFileExists($destination->fullPath());
-
-    // Assert the target contained the contents from the correct scaffold file.
-    $contents = trim(file_get_contents($destination->fullPath()));
-    $this->assertEquals('# Test version of robots.txt from drupal/core.', $contents);
+    // Assert that the target file was not created.
+    $this->assertFileNotExists($destination->fullPath());
 
     // Confirm that expected output was written to our io fixture.
     $output = $fixtures->getOutput();
-    $this->assertContains('Copy [web-root]/robots.txt from assets/robots.txt', $output);
+    $this->assertContains('Skip [web-root]/robots.txt: disabled', $output);
   }
 
 }
